requirements2.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
certifi==2025.7.14
cffi==1.17.1
click==8.1.8
cryptography==41.0.7
databases==0.9.0
elementpath==5.0.3
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
lxml==6.0.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyOpenSSL==23.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
xmlschema==4.1.0
